fix(search): switch to proper types for card stats
The StyleCard struct is using int64 for stats.  We might change that in
the future.  For now, this will resolve the template error within cards
on Search page.

+++ models/style.go
@@ -77,8 +77,8 @@ type StyleSearch struct {
 	Preview     string
 	DisplayName string
 	Username    string
-	Views       int
-	Installs    int
+	Views       int64
+	Installs    int64
 	User        User `gorm:"foreignKey:ID"`
 	UserID      uint
 }
